move the sql importer to an information source, add in a test information source
