+++ sites/rfh/settings.py
@@ -1,2 +1,3 @@
 DATABASE_STRING = 'postgresql://gloss:gloss@localhost/gloss_rfh'
+INFORMATION_SOURCE = 'sites.rfh.information_source.InformatinSourceOnTest'
 from sites.rfh.local_settings import *
